requirements.txt
Flask==2.3.2
flask-cors==4.0.0
requests==2.32.3
python-dotenv==1.0.1
Pillow==10.0.0